+++ src/tts/gather/api.ts
@@ -166,5 +166,6 @@ export class TextGather {
         await import('./tutorial-popup')
         await import('./mod-manager')
         await import('./equipment-menu')
+        await import('./quick-menu-food')
     }
 }